accessible/tests/mochitest/text/test_lineboundary.html

<!DOCTYPE html>
<html>
<head>
  <title>Line boundary getText* functions tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>
  <script type="application/javascript">
    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // __h__e__l__l__o__ __m__y__ __f__r__i__e__n__d__
      //  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15

      var IDs = [ "input", "div", "editable", "textarea",
                  getNode("ta", getNode("ta_cntr").contentDocument) ];

      testTextBeforeOffset(IDs, BOUNDARY_LINE_START,
                           [ [ 0, 15, "", 0, 0 ] ]);
      testTextBeforeOffset(IDs, BOUNDARY_LINE_END,
                           [ [ 0, 15, "", 0, 0 ] ]);

      testTextAtOffset(IDs, BOUNDARY_LINE_START,
                       [ [ 0, 15, "hello my friend", 0, 15 ] ]);
      testTextAtOffset(IDs, BOUNDARY_LINE_END,
                       [ [ 0, 15, "hello my friend", 0, 15 ] ]);

      testTextAfterOffset(IDs, BOUNDARY_LINE_START,
                          [ [ 0, 15, "", 15, 15 ] ]);
      testTextAfterOffset(IDs, BOUNDARY_LINE_END,
                          [ [ 0, 15, "", 15, 15 ] ]);

      //////////////////////////////////////////////////////////////////////////
      // __o__n__e__w__o__r__d__\n
      //  0  1  2  3  4  5  6  7
      // __\n
      //  8
      // __t__w__o__ __w__o__r__d__s__\n
      //  9 10 11 12 13 14 15 16 17 18

      IDs = [ "ml_div", "ml_divbr", "ml_editable", "ml_editablebr", "ml_textarea"];

      testTextBeforeOffset(IDs, BOUNDARY_LINE_START,
                           [ [ 0, 7, "", 0, 0 ],
                             [ 8, 8, "oneword\n", 0, 8 ],
                             [ 9, 18, "\n", 8, 9 ],
                             [ 19, 19, "two words\n", 9, 19 ]]);

      testTextBeforeOffset(IDs, BOUNDARY_LINE_END,
                          [ [ 0, 7, "", 0, 0 ],
                            [ 8, 8, "oneword", 0, 7 ],
                            [ 9, 18, "\n", 7, 8 ],
                            [ 19, 19, "\ntwo words", 8, 18 ]]);

      testTextAtOffset(IDs, BOUNDARY_LINE_START,
                       [ [ 0, 7, "oneword\n", 0, 8 ],
                         [ 8, 8, "\n", 8, 9 ],
                         [ 9, 18, "two words\n", 9, 19 ],
                         [ 19, 19, "", 19, 19 ]]);
      testTextAtOffset(IDs, BOUNDARY_LINE_END,
                       [ [ 0, 7, "oneword", 0, 7 ],
                         [ 8, 8, "\n", 7, 8 ],
                         [ 9, 18, "\ntwo words", 8, 18 ],
                         [ 19, 19, "\n", 18, 19 ]]);

      testTextAfterOffset(IDs, BOUNDARY_LINE_START,
                          [ [ 0, 7, "\n", 8, 9 ],
                            [ 8, 8, "two words\n", 9, 19 ],
                            [ 9, 19, "", 19, 19 ]]);
      testTextAfterOffset(IDs, BOUNDARY_LINE_END,
                          [ [ 0, 7, "\n", 7, 8 ],
                            [ 8, 8, "\ntwo words", 8, 18 ],
                            [ 9, 18, "\n", 18, 19 ],
                            [ 19, 19, "", 19, 19 ]]);

      //////////////////////////////////////////////////////////////////////////
      // a * b (* is embedded char for link)
      testTextBeforeOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_START,
                           [ [ 0, 5, "", 0, 0 ] ]);

      testTextBeforeOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_END,
                           [ [ 0, 5, "", 0, 0 ] ]);

      testTextAtOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_START,
                       [ [ 0, 5, "a " + kEmbedChar + " c", 0, 5 ] ]);

      testTextAtOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_END,
                       [ [ 0, 5, "a " + kEmbedChar + " c", 0, 5 ] ]);

      testTextAfterOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_START,
                          [ [ 0, 5, "", 5, 5 ] ]);

      testTextAfterOffset([ getAccessible("ht_1").firstChild ], BOUNDARY_LINE_END,
                          [ [ 0, 5, "", 5, 5 ] ]);

      //////////////////////////////////////////////////////////////////////////
      // foo<br> and foo<br><br>

      testTextAtOffset([ getAccessible("ht_2").firstChild.firstChild ],
                       BOUNDARY_LINE_START,
                       [ [ 0, 3, "foo", 0, 3 ] ]);
      testTextAtOffset([ getAccessible("ht_3").firstChild.firstChild ],
                       BOUNDARY_LINE_START,
                       [ [ 0, 3, "foo\n", 0, 4 ], [ 4, 4, "", 4, 4 ] ]);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="getTextAtOffset for word boundaries: beginning of a new life"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=853340">
    Bug 853340
  </a>
  <a target="_blank"
     title="getTextBeforeOffset for word boundaries: evolving"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=855732">
    Bug 855732
  </a>
  <a target="_blank"
     title=" getTextAfterOffset for line boundary on new rails"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=882292">
    Bug 882292
  </a>
  
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <input id="input" value="hello my friend"/>
  <div id="div">hello my friend</div>
  <div id="editable" contenteditable="true">hello my friend</div>
  <textarea id="textarea">hello my friend</textarea>
  <iframe id="ta_cntr"
          src="data:text/html,<html><body><textarea id='ta'>hello my friend</textarea></body></html>"></iframe>

  <pre>
    <div id="ml_div" style="border-style:outset;">oneword

two words
</div>
  <div id="ml_divbr" style="border-style:outset;">oneword<br/><br/>two words<br/><br/></div>
  <div id="ml_editable" style="border-style:outset;" contenteditable="true">oneword

two words
</div>
  <div id="ml_editablebr" contenteditable="true" style="border-style:outset;">oneword<br/><br/>two words<br/><br/></div>
  <textarea id="ml_textarea" cols="300">oneword

two words
</textarea>
  </pre>

  <iframe id="ht_1" src="data:text/html,<html><body>a <a href=''>b</a> c</body></html>"></iframe>

  <iframe id="ht_2" src="data:text/html,<div contentEditable='true'>foo<br/></div>"></iframe>
  <iframe id="ht_3" src="data:text/html,<div contentEditable='true'>foo<br/><br/></div>"></iframe>
</body>
</html>
